services/topic-modeling/requirements.txt
dspy-ai>=2.5.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.6.0
python-dotenv>=1.0.0
